chore: allow to use local cache in ec2 lambdas (#297)
* chore: allow to use local cache in ec2 lambdas

* feat: add support to use input use_dedicated_host

* chore: use branch feat-ec2-dynamic-config

* feat: use branch feat-macos-support

+++ modules/platform/ec2_deployment/scripts/download_lambdas.sh
@@ -9,12 +9,18 @@ fi
 DOWNLOAD_PATH="$1"
 VERSION="$2"
 
-rm -rf "$DOWNLOAD_PATH"
-mkdir -p "$DOWNLOAD_PATH"
+if [ -n "$USE_CACHE" ] && [ -n "$CACHE_PATH" ] && [ -d "$CACHE_PATH" ]; then
+    echo "USE_CACHE is set and $CACHE_PATH exists, skipping download." >&2
+    DOWNLOAD_PATH="$CACHE_PATH"
+    VERSION="local-cache"
+else
+    rm -rf "$DOWNLOAD_PATH"
+    mkdir -p "$DOWNLOAD_PATH"
 
-# Download files to the specified director
-wget --no-verbose -P "$DOWNLOAD_PATH" "https://github.com/github-aws-runners/terraform-aws-github-runner/releases/download/${VERSION}/runner-binaries-syncer.zip"
-wget --no-verbose -P "$DOWNLOAD_PATH" "https://github.com/github-aws-runners/terraform-aws-github-runner/releases/download/${VERSION}/runners.zip"
-wget --no-verbose -P "$DOWNLOAD_PATH" "https://github.com/github-aws-runners/terraform-aws-github-runner/releases/download/${VERSION}/webhook.zip"
+    # Download files to the specified directory
+    wget --no-verbose -P "$DOWNLOAD_PATH" "https://github.com/github-aws-runners/terraform-aws-github-runner/releases/download/${VERSION}/runner-binaries-syncer.zip"
+    wget --no-verbose -P "$DOWNLOAD_PATH" "https://github.com/github-aws-runners/terraform-aws-github-runner/releases/download/${VERSION}/runners.zip"
+    wget --no-verbose -P "$DOWNLOAD_PATH" "https://github.com/github-aws-runners/terraform-aws-github-runner/releases/download/${VERSION}/webhook.zip"
+fi
 
-echo -n "{\"version\":\"${VERSION}\"}"
+echo -n "{\"version\":\"${VERSION}\",\"path\":\"${DOWNLOAD_PATH}\"}"
